Rename `v1` package to `v1alpha7`
This commit makes the following changes:
- Renames the `v1` package to `v1alpha7` in the proto files.
- Renames the `v1` directory to `v1alpha7` in the proto files.
- Renames the `v1` directory to `v1alpha7` in the `py` directory.
- Rename the test file `test_common.py` to `test_common_v1alpha7.py`.
