--- proto/frequenz/api/common/v1/market/energy.proto
@@ -1,24 +1,0 @@
-// Frequenz definitions of energy.
-//
-// Copyright 2023 Frequenz Energy-as-a-Service GmbH
-//
-// Licensed under the MIT License (the "License");
-// you may not use this file except in compliance with the License.
-
-syntax = "proto3";
-
-package frequenz.api.common.v1.market;
-
-import "frequenz/api/common/v1/type/decimal.proto";
-
-// Represents a single unit of electricity.
-//
-// !!! note
-//     The unit of energy is denominated in MWh, which is a unit of energy
-//     representing total output over a period.(Megawatt-hours). This differs
-//     from MW (Megawatts), a unit of power representing the rate of energy
-//     production or consumption.
-message Energy {
-  // Energy unit in Megawatthours (MWh).
-  frequenz.api.common.v1.type.Decimal mwh = 1;
-}
