--- proto/frequenz/api/common/v1/microgrid/electrical_components/grid.proto
@@ -1,46 +1,0 @@
-// Frequenz definitions for grid connection points.
-//
-// Copyright:
-// Copyright 2023 Frequenz Energy-as-a-Service GmbH
-//
-// License:
-// MIT
-
-syntax = "proto3";
-
-package frequenz.api.common.v1.microgrid.electrical_components;
-
-// A representation of a grid connection point. This is the point where a
-// microgrid connects to the grid.
-//
-// The terms "Grid Connection Point" and "Point of Common Coupling" (PCC) are
-// commonly used in the context.
-//
-// While both terms describe a connection point to the grid, the
-// `GridConnectionPoint` is specifically the physical connection point of the
-// generation facility to the grid, often concerned with the technical and
-// ownership aspects of the connection.
-//
-// In contrast, the PCC is is more specific in terms of electrical engineering.
-// It refers to the point where a customer's local electrical system (such as a
-// microgrid) connects to the utility distribution grid in such a way that it
-// can affect other customers’ systems connected to the same network. It is the
-// point where the grid and customer's electrical systems interface and where
-// issues like power quality and supply regulations are assessed.
-//
-// The term `GridConnectionPoint` is used to make it clear that what is referred
-// to here is the physical connection point of the local facility to the grid.
-// Note that this may also be the PCC in some cases.
-message GridConnectionPoint {
-  // This refers to the maximum amount of electrical current, in amperes, that a
-  // fuse at the grid connection point is designed to safely carry under normal
-  // operating conditions.
-  //
-  // This limit applies to currents both flowing in or out of each of the 3
-  // phases individually.
-  //
-  // In other words, a current _i_ A at one of the phases of the grid connection
-  // point must comply with the following constraint:
-  // `-rated_fuse_current <= i <= rated_fuse_current`
-  uint32 rated_fuse_current = 1;
-}
