--- proto/frequenz/api/common/v1/microgrid/electrical_components/inverter.proto
@@ -1,36 +1,0 @@
-// Frequenz definitions for inverters.
-//
-// Copyright:
-// Copyright 2023 Frequenz Energy-as-a-Service GmbH
-//
-// License:
-// MIT
-
-syntax = "proto3";
-
-package frequenz.api.common.v1.microgrid.electrical_components;
-
-// Enumerated inverter types.
-enum InverterType {
-  // Unspecified.
-  INVERTER_TYPE_UNSPECIFIED = 0;
-
-  // Battery inverter.
-  INVERTER_TYPE_BATTERY = 1;
-
-  // PV (photovoltaic) inverter.
-  INVERTER_TYPE_PV = 2;
-
-  // Hybrid inverter.
-  INVERTER_TYPE_HYBRID = 3;
-
-  // Wind turbine inverter.
-  INVERTER_TYPE_WIND_TURBINE = 4;
-}
-
-// A representation of an inverter.
-// The inverter metadata.
-message Inverter {
-  // The inverter type.
-  InverterType type = 1;
-}
