--- proto/frequenz/api/common/v1/microgrid/electrical_components/transformer.proto
@@ -1,28 +1,0 @@
-// Frequenz definitions for transformers.
-//
-// Copyright:
-// Copyright 2023 Frequenz Energy-as-a-Service GmbH
-//
-// License:
-// MIT
-
-syntax = "proto3";
-
-package frequenz.api.common.v1.microgrid.electrical_components;
-
-// A representation of a voltage transformer.
-// Voltage transformers are used to step up or step down the voltage, keeping
-// the power somewhat constant by increasing or decreasing the current.
-// If voltage is stepped up, current is stepped down, and vice versa.
-// Note that voltage transformers have efficiency losses, so the output power
-// is always less than the input power.
-message VoltageTransformer {
-  // The primary voltage of the transformer.
-  // This is the input voltage that is stepped up or down.
-  float primary = 1;
-
-  // The secondary voltage of the transformer.
-  // This is the output voltage that is the result of stepping the primary
-  // voltage up or down.
-  float secondary = 2;
-}
