+++ py/frequenz/api/common/v1alpha7/__init__.py
@@ -1,0 +1,7 @@
+# License: MIT
+# Copyright © 2023 Frequenz Energy-as-a-Service GmbH
+
+"""Frequenz common gRPC API and bindings.
+
+TODO(cookiecutter): Add a more descriptive module description.
+"""
